# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Event Bot API for managing events and scheduling.",
        "title": "Go Event Bot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1/",
    "paths": {
        "/events": {
            "get": {
                "description": "List all upcoming events",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "List Events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event.Instance.Get"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Create an event",
                "parameters": [
                    {
                        "description": "create an event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event.Create.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "Get event information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get Event Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Event.Get"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Update Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update an event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event.Create.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Delete Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "description": "List all tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "List Tags",
                "responses": {
                    "200": {
                        "description": "tag list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Tag.Get"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Create Tag",
                "parameters": [
                    {
                        "description": "required body for creating a new tag",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tag.Create.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "tag information",
                        "schema": {
                            "$ref": "#/definitions/Tag.Get"
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "description": "Get tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Get Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the tag",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag information",
                        "schema": {
                            "$ref": "#/definitions/Tag.Get"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Update Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the tag",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body to update the tag",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tag.Update.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag information",
                        "schema": {
                            "$ref": "#/definitions/Tag.Get"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Delete Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the tag",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "user signup",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user account",
                "parameters": [
                    {
                        "description": "body to register",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.Register.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user information",
                        "schema": {
                            "$ref": "#/definitions/data.User"
                        }
                    }
                }
            }
        },
        "/users/activated": {
            "post": {
                "description": "Activate a new user account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Activate a new user account",
                "parameters": [
                    {
                        "description": "activation token",
                        "name": "activation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.Activation.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user information",
                        "schema": {
                            "$ref": "#/definitions/data.User"
                        }
                    }
                }
            }
        },
        "/users/password": {
            "post": {
                "description": "Change current user password",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update Password",
                "parameters": [
                    {
                        "description": "body to register",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.UpdatePassword.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag information",
                        "schema": {
                            "$ref": "#/definitions/data.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Event.Create.Request": {
            "description": "Information required to update a new event",
            "type": "object",
            "properties": {
                "channel_id": {
                    "description": "Discord ID of the channel the bot will post the event",
                    "type": "integer"
                },
                "description": {
                    "description": "Description/Content of the event",
                    "type": "string"
                },
                "duration": {
                    "description": "Duration of the each event following ISO 8601 standard",
                    "type": "string",
                    "example": "PT30M"
                },
                "guild_id": {
                    "description": "Discord Guild ID/server the bot will publish on",
                    "type": "integer"
                },
                "is_active": {
                    "description": "If the event is active of not",
                    "type": "boolean"
                },
                "rrule": {
                    "description": "Recurrence rule following RFC 5545 https://icalendar.org/rrule-tool.html",
                    "type": "string",
                    "example": "FREQ=WEEKLY;INTERVAL=1;BYDAY=MO;UNTIL=20250731T000000Z"
                },
                "title": {
                    "description": "Title of the  event",
                    "type": "string"
                }
            }
        },
        "Event.Get": {
            "description": "Event Response object",
            "type": "object",
            "properties": {
                "channel_id": {
                    "description": "Discord ID of the channel the bot will post the event",
                    "type": "integer"
                },
                "created_date": {
                    "description": "Creation date of the event",
                    "type": "string"
                },
                "description": {
                    "description": "Description/Content of the event",
                    "type": "string"
                },
                "duration": {
                    "description": "Duration of the event following the ISO 8601 standard",
                    "type": "string",
                    "example": "PT30M"
                },
                "guild_id": {
                    "description": "Discord Guild ID/server the bot will publish on",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the event",
                    "type": "string"
                },
                "is_active": {
                    "description": "Tell if the event is currently active",
                    "type": "boolean"
                },
                "rrule": {
                    "description": "Recurrence rule following RFC 5545 https://icalendar.org/rrule-tool.html",
                    "type": "string",
                    "example": "FREQ=WEEKLY;INTERVAL=1;BYDAY=MO;UNTIL=20250731T000000Z"
                },
                "title": {
                    "description": "Title of the event",
                    "type": "string"
                },
                "updated_date": {
                    "description": "Last modification of the event",
                    "type": "string"
                }
            }
        },
        "Event.Instance.Get": {
            "description": "Event Instance Response object",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the event",
                    "type": "string"
                },
                "duration": {
                    "description": "Duration of the event following the ISO 8601 standard",
                    "type": "string",
                    "example": "PT30M"
                },
                "end_date": {
                    "description": "End date of the event",
                    "type": "string"
                },
                "event_id": {
                    "description": "ID of the event",
                    "type": "string"
                },
                "start_date": {
                    "description": "Start date of the event",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the event",
                    "type": "string"
                }
            }
        },
        "Tag.Create.Request": {
            "description": "Create Tag Model",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Short description of the tag",
                    "type": "string",
                    "example": "regroup events for warriors"
                },
                "name": {
                    "description": "Name of the tag",
                    "type": "string",
                    "example": "Warriors"
                }
            }
        },
        "Tag.Get": {
            "description": "Tag information",
            "type": "object",
            "properties": {
                "created_date": {
                    "description": "Creation date of the tag",
                    "type": "string"
                },
                "description": {
                    "description": "Description of the tag",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the tag",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the tag",
                    "type": "string"
                },
                "updated_date": {
                    "description": "Last modification on the tag",
                    "type": "string"
                }
            }
        },
        "Tag.Update.Request": {
            "description": "Request body to update the tag",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "User.Activation.Request": {
            "description": "Activate token request",
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token",
                    "type": "string"
                }
            }
        },
        "User.Register.Request": {
            "description": "User Registration Body",
            "type": "object",
            "properties": {
                "email": {
                    "description": "email of the user",
                    "type": "string"
                },
                "name": {
                    "description": "username of the user",
                    "type": "string"
                },
                "password": {
                    "description": "password",
                    "type": "string"
                }
            }
        },
        "User.UpdatePassword.Request": {
            "description": "User Request to update password",
            "type": "object",
            "properties": {
                "password": {
                    "description": "User password",
                    "type": "string"
                },
                "token": {
                    "description": "Authorization token",
                    "type": "string"
                }
            }
        },
        "data.User": {
            "type": "object",
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}